python/category/requirements.txt
aioconsole==0.7.1
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.3.0
attrs==23.2.0
backoff==2.2.1
black==24.4.2
botocore==1.34.106
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
frozenlist==1.4.1
gql==3.5.0
graphql-core==3.2.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jmespath==1.0.1
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.1
urllib3==2.2.1
websockets==11.0.3
yarl==1.9.4
